+++ app/controllers/app.js
@@ -40,6 +40,7 @@ var express = require('express')
     , LRU         = require('lru-cache')
     , formats     = require(global.settings.app_root + '/app/models/formats')
     , HealthCheck = require(global.settings.app_root + '/app/monitoring/health_check')
+    , PgErrorHandler = require(global.settings.app_root + '/app/postgresql/error_handler')
     ;
 
 var metadataBackend = MetadataDB({
@@ -486,6 +487,8 @@ function handleQuery(req, res) {
                 formatter.sendResponse(opts, this);
             },
             function errorHandle(err){
+                formatter = null;
+
                 if ( err ) handleException(err, res);
                 if ( req.profiler ) {
                   req.profiler.sendStats(); // TODO: do on nextTick ?
@@ -559,10 +562,19 @@ function generateMD5(data){
 }
 
 
-function handleException(err, res){
-    var msg = (global.settings.environment == 'development')
-        ? {error:[err.message], stack: err.stack}
-        : {error:[err.message]};
+function handleException(err, res) {
+    var pgErrorHandler = new PgErrorHandler(err);
+
+    var msg = {
+        error: [pgErrorHandler.getMessage()]
+    };
+
+    _.defaults(msg, pgErrorHandler.getFields());
+
+    if (global.settings.environment == 'development') {
+        msg.stack = err.stack;
+    }
+
     if (global.settings.environment !== 'test'){
         // TODO: email this Exception report
         console.error("EXCEPTION REPORT: " + err.stack)
@@ -579,19 +591,16 @@ function handleException(err, res){
       res.header('X-SQLAPI-Profiler', res.req.profiler.toJSONString());
     }
 
-    res.send(msg, getStatusError(err, res.req));
+    res.send(msg, getStatusError(pgErrorHandler, res.req));
 
     if ( res.req && res.req.profiler ) {
       res.req.profiler.sendStats();
     }
 }
 
-function getStatusError(err, req) {
-    var statusError = _.isUndefined(err.http_status) ? 400 : err.http_status;
+function getStatusError(pgErrorHandler, req) {
 
-    if (err.message && err.message.match(/permission denied/)) {
-        statusError = 401;
-    }
+    var statusError = pgErrorHandler.getStatus();
 
     // JSONP has to return 200 status error
     if (req && req.query && req.query.callback) {